pyproject.toml
[project]
name = "slumworld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "click==8.1.8",
    "geopandas==0.13.2",
    "imageio==2.35.1",
    "matplotlib==3.7.5",
    "numpy>=1.24.4",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pandas==2.0.3",
    "pillow==10.4.0",
    "pydensecrf",
    "pytorch-lightning==1.9.0",
    "pytorch-toolbelt==0.8.0",
    "pyyaml==6.0.2",
    "rasterio==1.3.11",
    "ruamel-yaml==0.18.10",
    "ruamel-yaml-clib==0.2.8",
    "scikit-image==0.21.0",
    "scikit-learn==1.3.2",
    "tabulate==0.9.0",
    "tensorboard==2.14.0",
    "timm>=1.0.19",
    "torchmetrics==0.11.4",
    "tqdm==4.67.1",
    "transformers==4.28.0",
]

[tool.uv.sources]
pydensecrf = { git = "https://github.com/lucasb-eyer/pydensecrf.git" }
